LayoutTests/platform/chromium-linux/plugins/mouse-events-fixedpos-expected.txt

CONSOLE MESSAGE: line 39: PLUGIN: getFocusEvent
CONSOLE MESSAGE: line 39: PLUGIN: mouseDown at (50, 50)
CONSOLE MESSAGE: line 40: PLUGIN: mouseUp at (50, 50)
CONSOLE MESSAGE: line 42: PLUGIN: mouseDown at (60, 60)
CONSOLE MESSAGE: line 44: PLUGIN: mouseUp at (70, 60)

Tests for widget positions being correctly updated after scrolling. rdar://problem/7559069
